SEQNG-978 Refactoring of common code suggested on PR
