+++ modules/acm/src/test/scala/edu/gemini/epics/acm/ObserveStateSpec.scala
@@ -23,7 +23,10 @@ import org.scalatest._
   * care about the state of the channels. Instead we want to only observe
   * the state transitions
   */
-@SuppressWarnings(Array("org.wartremover.warts.NonUnitStatements", "org.wartremover.warts.PublicInference", "org.wartremover.warts.IsInstanceOf", "org.wartremover.warts.AsInstanceOf"))
+@SuppressWarnings(
+  Array("org.wartremover.warts.NonUnitStatements",
+        "org.wartremover.warts.PublicInference",
+        "org.wartremover.warts.IsInstanceOf"))
 final class ObserveStateSpec extends FunSuite with MockFactory {
 
   test("NIFS normal observation") {
@@ -923,6 +926,7 @@ final class ObserveStateSpec extends FunSuite with MockFactory {
 // This is exactly the case for `ReadWriteClientEpicsChannel[T]` and the [register|unRegister]Listener methods
 // The workaround is to create classes such as below with concrete type and mock those
 // https://github.com/paulbutcher/ScalaMock/issues/193
+@SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
 class CadDirectiveChannelMock extends ReadWriteClientEpicsChannel[CadDirective] {
   override def setValue(a: java.util.List[CadDirective]): Unit = {}
   override def setValue(a : CadDirective): Unit = {}
@@ -939,6 +943,7 @@ class CadDirectiveChannelMock extends ReadWriteClientEpicsChannel[CadDirective] 
   override def destroy(): Unit = {}
 }
 
+@SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
 class CarStateChannelMock extends ReadWriteClientEpicsChannel[CarState] {
   override def setValue(a: java.util.List[CarState]): Unit = {}
   override def setValue(a : CarState): Unit = {}
@@ -955,6 +960,7 @@ class CarStateChannelMock extends ReadWriteClientEpicsChannel[CarState] {
   override def destroy(): Unit = {}
 }
 
+@SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
 class IntChannelMock extends ReadWriteClientEpicsChannel[JInteger] {
   override def setValue(a: java.util.List[JInteger]): Unit = {}
   override def setValue(a : JInteger): Unit = {}
@@ -971,6 +977,7 @@ class IntChannelMock extends ReadWriteClientEpicsChannel[JInteger] {
   override def destroy(): Unit = {}
 }
 
+@SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
 class ShortChannelMock extends ReadWriteClientEpicsChannel[JShort] {
   override def setValue(a: java.util.List[JShort]): Unit = {}
   override def setValue(a : JShort): Unit = {}
@@ -987,6 +994,7 @@ class ShortChannelMock extends ReadWriteClientEpicsChannel[JShort] {
   override def destroy(): Unit = {}
 }
 
+@SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
 class StringChannelMock extends ReadWriteClientEpicsChannel[String] {
   override def setValue(a: java.util.List[String]): Unit = {}
   override def setValue(a : String): Unit = {}
